updated contact.html and stylessheets
removed a styling class from my button tag in contact html, and deleted its ruleset in the css files.

added a transition effect on my h1 tags and header container items , in stylesheet.

updated the styling of my button tag and button class, on stylesheet.

updated the transition effect on the pseudo classes of the button/button class, in stylesheet.

added a transition effect on the portrait class, in stylesheet.

added a border on .active class- and removed background color, in stylesheet.

commented the styling on the  .standard label class, in stylesheet.

added indentation at the bottom of stylesheet.

updated the production file by autoprefixing the .css file